backend/tests/conftest.py
import os
import sys

import pytest
from dotenv import load_dotenv
from fastapi.testclient import TestClient

# Add parent directory to path for imports
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

# Load environment variables from .env file
load_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), ".env"))

from app.main import app
from app.chat import ChatService


@pytest.fixture
def client():
    """FastAPI test client."""
    return TestClient(app)


@pytest.fixture
def chat_service():
    """ChatService instance for testing."""
    return ChatService()
