backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
openai>=1.50.0
numpy==2.1.0
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
